Updating to 1.19.4 Part 1
Also fixed a bug with configs that I should backport eventually. Currently, the crafting and camera roll are still broken.

+++ src/main/java/me/appw/vikare/core/network/client/CPlayerFlappingPacket.java
@@ -42,7 +42,7 @@ public class CPlayerFlappingPacket {
                 slotOpt.ifPresent(slotResult -> {
                     ItemStack wings = slotResult.stack();
                     if (!sender.isCreative() && !wings.is(WingItem.FREE_FLIGHT)) {
-                        sender.causeFoodExhaustion(VikareConfig.COMMON.exhaustionAmount.get());
+                        sender.causeFoodExhaustion(VikareConfig.COMMON.exhaustionAmount.get().floatValue());
                     }
 
                     if (message.state != FlappingState.NONE) {
